fix: created job form parent page with susepnse boundary
